[REFACTOR] #3 디자인 시스템 button component 리펙토링
