+++ core/ui/src/main/java/pokitmons/pokit/core/ui/components/atom/button/attributes/ButtonAttributes.kt
@@ -4,7 +4,7 @@ enum class PokitButtonShape {
     ROUND, RECTANGLE,
 }
 
-enum class PokitButtonState {
+internal enum class PokitButtonState {
     DEFAULT, DISABLE,
 }
